shell: Use `identity.follows=""` (meaning `self`)
The identity input is used to override the user details for Git and
other programs that I pre-configure with my personal credentials. I
recently learned that you can use `.follows = ""` to create references
to self in your inputs. This removes the need for a separate repo to
store the identity information.