shell: Wrap `gh` for single-user auth
Say you are logged in to 2 users: alice and bob. When you `nix develop`
in one project, you want to always be alice. On another project you
always want to be bob. This wrapper will hide the other user (even
though the auth token is still in the keyring) from `gh`. Combined with
the `git` wrapper, you get a consistent GitHub identity.